moved the hashmap into ONE MainActivity
and making a new branch to keep it that way
